requirements.txt
python-binance==1.0.19
typer==0.12.3
rich==13.7.1
python-dotenv==1.0.1